feat: You can now enter translations when adding a word
Translations (in the two other languages) can now be entered when adding a new word.
Added Language enum and a languageToString() function for printing.
The program now remembers the language choice and the two other languages in a Context structure and passes it to command functions to use. All language prints are now aware of the the language choice. 😎